SiegeChatbot/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "cohere>=5.16.3",
    "python-dotenv>=1.1.1",
    "python-telegram-bot>=22.3",
]
